+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.15.0
+
+### Added
+
+- `project-loader` will now load project libraries from `.mps/libraries.xml` under the project directory when using MPS
+  environment (can be disabled via `--no-libraries`).
+
 ## 1.14.3
 
 ### Fixed